requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
gradio==4.19.0

# ML & Data
scikit-surprise==1.1.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0

# Evaluation & Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Utilities
joblib==1.3.0
python-dotenv==1.0.0
